+++ sdk/cognitivelanguage/azure-ai-language-conversations/samples/sample_analyze_orchestration_app_qna_response.py
@@ -28,14 +28,7 @@ def sample_analyze_orchestration_app_qna_response():
     # import libraries
     import os
     from azure.core.credentials import AzureKeyCredential
-
     from azure.ai.language.conversations import ConversationAnalysisClient
-    from azure.ai.language.conversations.models import (
-        CustomConversationalTask,
-        ConversationAnalysisOptions,
-        CustomConversationTaskParameters,
-        TextConversationItem
-    )
 
     # get secrets
     clu_endpoint = os.environ["AZURE_CONVERSATIONS_ENDPOINT"]
@@ -48,36 +41,43 @@ def sample_analyze_orchestration_app_qna_response():
     with client:
         query = "How are you?"
         result = client.analyze_conversation(
-                task=CustomConversationalTask(
-                    analysis_input=ConversationAnalysisOptions(
-                        conversation_item=TextConversationItem(
-                            text=query
-                        )
-                    ),
-                    parameters=CustomConversationTaskParameters(
-                        project_name=project_name,
-                        deployment_name=deployment_name
-                    )
-                )
-            )
+            task={
+                "kind": "Conversation",
+                "analysisInput": {
+                    "conversationItem": {
+                        "participantId": "1",
+                        "id": "1",
+                        "modality": "text",
+                        "language": "en",
+                        "text": query
+                    },
+                    "isLoggingEnabled": False
+                },
+                "parameters": {
+                    "projectName": project_name,
+                    "deploymentName": deployment_name,
+                    "verbose": True
+                }
+            }
+        )
 
     # view result
-    print("query: {}".format(result.results.query))
-    print("project kind: {}\n".format(result.results.prediction.project_kind))
+    print("query: {}".format(result["result"]["query"]))
+    print("project kind: {}\n".format(result["result"]["prediction"]["projectKind"]))
 
     # top intent
-    top_intent = result.results.prediction.top_intent
+    top_intent = result["result"]["prediction"]["topIntent"]
     print("top intent: {}".format(top_intent))
-    top_intent_object = result.results.prediction.intents[top_intent]
-    print("confidence score: {}".format(top_intent_object.confidence))
-    print("project kind: {}".format(top_intent_object.target_kind))
+    top_intent_object = result["result"]["prediction"]["intents"][top_intent]
+    print("confidence score: {}".format(top_intent_object["confidenceScore"]))
+    print("project kind: {}".format(top_intent_object["targetProjectKind"]))
 
-    if top_intent_object.target_kind == "question_answering":
+    if top_intent_object["targetProjectKind"] == "QuestionAnswering":
         print("\nview qna result:")
-        qna_result = top_intent_object.result
-        for answer in qna_result.answers:
-            print("\nanswer: {}".format(answer.answer))
-            print("answer: {}".format(answer.confidence))
+        qna_result = top_intent_object["result"]
+        for answer in qna_result["answers"]:
+            print("\nanswer: {}".format(answer["answer"]))
+            print("answer: {}".format(answer["confidenceScore"]))
 
     # [END analyze_orchestration_app_qna_response]