+++ sdk/cognitivelanguage/azure-ai-language-conversations/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
 
+## 1.1.0b1 (Unreleased)
+
+### Features Added
+* Conversation issue summarization task (Long-running operation)
+* Conversation PII extraction task (Long-running operation)
+
+### Breaking Changes
+* Client now uses python dictionaries for method parameters and results instead of classes.
+* Many input and result parameter name changes in `analyze_conversation()` method
+
+
 ## 1.0.0b3 (2022-04-19)
 
 ### Features Added